requirements.txt
pandas
openpyxl
numpy
tqdm
torch
sentence-transformers==2.7.0
transformers==4.40.2
tokenizers==0.19.1
huggingface-hub==0.36.0
accelerate>=0.29.0
safetensors>=0.4.3
google-genai
python-dotenv
